Consolidate query_companies views under main_user blueprint
Before this change we had two separate views for querying the
companies registered with the arbeitszeitapp, one for members and one
for companies. After this change we will one single view/route
"/user/query_companies" that provides the exact same functionality as
the two routes that we had before this change.

The rationale behind the change is that any authenticated user should
have access to this basic query functionality and thus the route
should be located under the `main_user` blueprint.

The integration tests for the query_companies view were slightly
adjusted for better coverage.