Add html for company's review consumptions feature
This commit adds the html layer for the company's review consumptions feature. This view allows companies to see a list of their "sales", with the buyer's non-anonymized data. If the consumer is a member, only the user's name is shown - in the long run this might not be detailed enough and we probably will have to identify the member with a link to their profile (currently there are no member profile pages that can be accessed by other users) or the member's ID number.