Anonymize worker data in prd account (business level)
Anonymize data of consumers that appears in the prd account of companies, if the consumer is a member (worker). This was done in the business logic, i.e. in the `ShowPRDAccountDetailsUseCase`.

I also changed some naming: In that use case all peers were named "Buyer", even if it was Social Accounting that was granting a "credit", so to speak. The name buyer was replaced by peer.

Also a view flask integration tests have been added, as there were none yet.