Add strict undefined behavior for Jinja2 environment
In the past our app did crash or behave unexpectedly because of errors in our jinja templates. For example when we would loop over undefined variables or use a mistyped variable in a template.  A current example is described in issue #921.

In order to be able to catch these errors in our tests we set the `undefined` paramter of the jinja environment of this app to `StrictUndefined` https://jinja.palletsprojects.com/en/3.0.x/api/#undefined-types

We also fix the places where the new configuration raises errors. In the `configuration_base.py` we have to insert four dummy config entries until a bug in flask-restx is fixed.