Fix bug in email confirmation timestamp
Before this change we had a bug when it came to confirming member
email addresses. The bug was that we would set the email confirmation
timestamp for confirmed members unconditionally to datetime.min. This
change addresses this bug and sets it to the actual time of
confirmation according to the DatetimeService implementation (which
relies on datetime.datetime.now).

We also extended the response model for the
GetUserAccountDetailsUseCase to show the time of confirmation for a
users email address. This was done to create test for the bug
addressed here. In the future this extension could be used to show the
email confirmation timestamp to users.