API: Refactor input doc and document path params
Before this change we used the `flask_restx.namespace.expect` decorator to document the expected inputs of our routes. Now we use a custom `with_input_documentation`decorator that uses internally a call to `flask_restx.namespace.doc`. This makes it clearer that we are only documenting and not (yet) parsing or validating the input with these expected input classes. The new implementation allows us now also to document path parameters like in `plans/<plan_id>`.