my_cooperations view: Use Notifier interface
Before this commit the my_cooperations route for companies used a
custom system for displaying user notifications in the web
frontend. This change modifies the responsible presenter and view
template such that the standard notification sytem
`arbeitszeit_web.notification.Notifier` is used instead. Because of
this change we in the presenter there is no need to handle user
notifications explicitly in view template. Instead we can rely on the
notification macro used shared by all views that inherit from
`base.html`.

The parameters for the ShowMyCooperationsPresenter.present method were
made keyword parameters. This change allowed us to make some of those
parameters optional. Although this is not very useful for the
production code it is however useful to improve the clarity of the
test cases. Making `accept_cooperation_response`,
`deny_cooperation_response` and
`cancel_cooperation_solicitation_response` optional allowed us to not
specify those arguments in test cases where we were not interested in
the corresponding behavior. This will also be useful in the future,
when we overhaul the my_cooperations_view such that there are separate
request handlers for the `AcceptCooperation`, `RejectCooperation` and
`CancelCooperation` use case.