Move parts of tile generation from company dashboard to presenter
Before this change, the presenter of the company dashboard did not contain any information on the tiles. This information was more or less hardcoded in the `dashboard.html` file. With this commit the presenter provides at least the info of the "accounts" tile.

It might be problematic that the `Tile.icon` attribute (in the abstract web layer)  contains references to "font awesome"-specific icon strings.