Remove cooperation data from Plan record
Before this change we stored the `cooperation` that a plan was part as
a field of the `Plan` record. This was also reflected in the database
implementation where `cooperation` was a column in the `plan`
table. After this change the `cooperation` attribute is removed from
the `Plan` record. Also the DB implementation stores the cooperation
data in a new separate table called `plan_cooperation`. The DB
interface had to changed slightly because of this, which affected some
use cases in a minor way.