Add info on P account page
In an effort to demystify the "balance" info on the company's p account page, two more pieces of information have been added: "sum of planned p" and "sum of consumed p". The "balance" is then apparently nothing else than the consumed hours subtracted from the planned hours.

In the `ShowPAccountDetailsUseCase` there is now one database request less needed.

If you find this change sensible I would apply this pattern also to the R, A and PRD accounts.