Introduce JsonList to possible json schema definition types
Before this commit the schema definition for json endpoints had
embedded the concept of a list in json into all the other schema types
by having the attribute `as_list`. This attribute was completely
removed. Instead we introduced a decicated type `JsonList` that
describes the fact that a certain value should be a list of
values. Unfortunately this change broke all the schema tests in our
API code that relied on the `as_list` feature. All schema tests that
broke were removed without a replacement. We did not bother with a
replacement because it is clear the API subsystem needs a complete
re-design as it relies on deprecated libraries, an awkward interface
an a schema definition system that is completely decoupled from the
values that are actually produced by the API endpoints.