Regenerate type stubs
After the latest update of `nixpkgs` we need to create the type stubs
for our python dependencies again.  This is achieved by this
commit. Further the type_stubs are now subject to automatic code
formatting to avoid unnecessary large diffs should mypy change the
code formatting of the generated type stubs for some reason.