Separate html templates for draft editing and draft creation
Before this commit we use the same html template for draft creation
and draft editing.  This was problematic because those pages are
different in their behavior and context information they need to
render properly. With this commit we created a separate html template
for editing a plan draft. This new template is a copy of the old
template with minor modifications to the displayed text.