Simplify html template for draft creation
Before this commit the web logic around draft creation was slightly
complicated. The complication was that the "Cancel" action in the
creation form was handled as a POST request that would simply redirect
the user to another page. Now the same thing is achieved via a `<a>`
element. This allowed us to simplify the route/view.