Update integration tests to use MockEmailService instead of flask_mail.
Previously, we used `flask_mail` for testing email notifications. Now,
we've created `MockEmailService`, designed with a similar interface to
the testing interface of `flask_mail`, and integrated it with the new
`MailPlugin`. This change makes our flask integration tests no longer
dependent on `flask_mail`, paving the way to remove flask_mail from
our codebase in the future.