Create ChangeUserEmailAddressController
This controller is intended to be used in the `change_email_address` route in `arbeitszeit_flask/user/routes.py`. It has two methods. One, `extract_email_addresses_from_token` is used in the GET and the POST method of the route, the second method, `create_use_case_request`, is used in the POST method and creates the UseCaseRequest object of the `ChangeUserEmailAddressUseCase`.