Add ListRegisteredHoursWorkedController
This commit adds the `ListRegisteredHoursWorkedController` class, which is responsible for creating a request object for the `list_registered_hours_worked` use case. The controller retrieves the current user role and company from the session and asserts that the current user role is `company`. The company ID is then passed as a parameter to the request object.

See issue #1029 for context.