Add ListRegisteredHoursWorkedPresenter
This commit adds the `ListRegisteredHoursWorkedPresenter` class, which is responsible for presenting the response of the `list_registered_hours_worked` use case. The presenter formats the registered hours worked data, including rounding decimal hours to two decimals and converting worker IDs to strings. This presenter will be used in the web layer to display the registered hours worked information.

See issue #1029 for context.