Check password in RequestEmailAddressChangeUseCase
With this commit, we are introducing a password check in the `RequestEmailAddressChangeUseCase`. It checks if the `current_password` value from the `Request` object is matching the password of the user account belonging to the `current_email_address` from the `Request` object.

Moreover, to be able to communicate the possible rejection reasons of the use case further down the line, the use case response returns now an instance of `RejectionReason` if rejected.