testing: Allow for freezing and advancing time with arbitrary values
This commit allows test authors to freeze time at arbitrary
dates. This is useful when you want to ensure that test steps should
be performed in a distinct order but you don't care about the specific
point in time when they happen. Similarly we also allow the
advancement of time in arbitrary quantities now.