Implement dedicated recording of hours worked
This PR implements the dedicated recording of registered hours
worked. Previously we only recorded hours worked via account
transaction from the labour account of a company to the individual
account of a worker.  This is in fact currently sufficient to
reconstruct the hours worked of every single member in the
system. This changed was made because it is necessary to move away
from the notion that a single registration of hours worked would
result in a single transaction in the public ledger but instead would
need to split up into several transactions in the future. To elaborate
one such transaction for every hour worked would be the portion of
labour that is allocated for public plans. In the future we would most
certainly want to create a separate transaction that would go into the
public sector funds. Other future transactions are to be expected as
well if we look towards introducing the notion of accumulation, where
an additional portion of every registered work hour would go towards
some accumulation funds instead of going into individual consumption.

This commit creates a new table in the DB to serve that end. The tests
in the business logic side of things were not subject to change as the
behavior of the application should not change in a way observable to
the users.