Provide minified bulma.css version
This commit changes the static file
`arbeitszeit_flask/static/bulma.css`. Before we used the regular
bulma.css file in our application. With this change we provide instead
a "minified" version of the same file. This version of bulma.css which
provides the same functionality as the old version is 80 kB
smaller. This should provide a significant reduction in network
traffic for setups where static files are not compressed for
transmission.