Create dedicated flask route for AcceptCooperationRequest use case
This PR separates the web logic for the AcceptCooperationRequest use
case from the company.my_cooperations routes.  This separation removes
the interdepence of the AcceptCooperationRequest web logic with that
of the use cases ListCoordinationsOfCompany, DenyCooperation,
ListMyCooperatingPlansUseCase and CancelCooperationSolicitation. All
existing tests where moved to maintain the same level of test
coverage. Also a redimentary test case for the flask routes was
implemented.