Isolate web logic for DenyCooperation use case
Before this commit the web logic for the DenyCooperation was tied up
with the `my_cooperations` route. This was undesirable and thusly the
web logic was isolated into its own route, view and presenter. The
HTML form on the `my_cooperations` route needed to be adjusted
slightly. Also a new very basic test for the new route was implemented
that checks that the web server at least does not crash when
submitting a post request against it.