Add integration test of icon path
This commit adds an integration test that checks if the expected SVG path of the "create new plan icon", as specified in `arbeitszeit_flask/templates/icons/file-circle-plus.html`, appears  on the company dashboard.

I am aware that this test is very specific, but in no other place we test for the concrete icon path value.