Fix bug in CreateDraftView
Before this change there was a bug in CreateDraftView. The was long
undetected as it is hard to trigger the failure resulting from the
bug. The user would need to submit form data to the view that
contained erroneous planning data. This would have led to an exception
being raised in a CreatePlanDraft use case object. The application
would have "crashed" with a 500 error response instead of rendering
the original form back to the user with an error message. The bug was
addressed by including proper form validation in POST handler of the
method and returning the rendered form to the user when form
validation fails.