Remove NixOS 23.11 from flake inputs
This commit removes the definitions for NixOS 23.11 from the flake
inputs of this repository.  This was done to reduce build and test
times. Since all systems that (are currently known to) run the
arbeitszeitapp are on NixOS 24.05 there is no practical need for
testing with the older stable version of NixOS.