Changer error propagation in `icon` template filter
This change concerns the error propagation inside the `icon` template
filter.  Before this change some exception raise as part of the
`icon_filter` routine were changed to include the icon name as part of
the exception message.  After this change the message of those
exceptions will be left as is and instead the `add_note` method is
used to add further context to the exception. The `add_note` method is
new in Python 3.11. See
https://docs.python.org/3.11/library/exceptions.html#BaseException.add_note
for more information on the topic.

Also the error message slightly changed to indicate that exceptions
occurred during the loading process of a specific icon.