Implement PreviewPrivateConsumptionUseCase
This PreviewPrivateConsumptionUseCase implemented in this commit is
supposed to be included in the workflow for members to register their
individual consumption of goods and services through the app.  The use
case is not yet hooked into the web code of the app. If hooked into
the web code this use case would provide basic product information for
a given plan id and the total consumption cost in case the amount is
also provided. The information will be displayed for the user so that
they can more easily identify what kind of product and how much they
are about to "consume".