Reimplement register_private_consumption web logic
This commit reimplements the web logic for registering a private
consumption in the arbeitszeitapp. Previously our approach relied
heavily on wtforms. This approarch proved inflexible since it is very
hard to manipulate field properties of wtforms forms reliably and
safely. Hence, we reimplemented the web logic with standard templating
techniques where the forms data in encoded in a basic
"dataclass".

This change is made in preparation for an upcoming change where the
individual fields of the form in question can be set to "readonly"
conditionally.