Allow QueryString objects to yield multiple values for one key
This commit changes the interface to `QueryString` objects. Before
this change `QueryString` objects were only able to yield a single
value for a given key. Now the method `QueryString.get` returns a list
of all values supplied for a given key. The `get_last_value` method
was added for convenient access to (possible) single value for cases
where the app does not need to handle multiple for a single key.

This change was made because the old interface to the query string
part of a URI was overly simplistic. It hid the fact that a query
string can have multiple values for a single key from the
developer. The new interface makes easier to remember the fact that a
query string does not need to have only a single value for a key and
makes it explicit that we choose a particular value from a possible
set of values, which is currently always the last specified value in a
query string.