Allow for handling all possible json values via Request interface
This commit changes the interface to `Request` objects with respect to
request body parsing slightly. This affects the `get_json` method
only.

Before the `get_json` method would have the signature `get_json(str)
-> str | None`. This signature only allows for json values to be
handled that are objects with strings as values. This is only a very
narrow subset of possible json values.

We introduced a type `JsonValue` that models all possible json values
after parsing. Json objects are mapped to `dict` and json arrays to
`list`. Numbers are mapped to either `int` or `float`. The `get_json`
method now returns an object of that type or `None`.