Refactor QueryCompaniesController
Before this commit the QueryCompaniesController would receive http
request object its supposed to process via its contructor. After this
change it receives the request object via an argument to
`import_form_data`. This makes the implementation more flexible since
it does not tie the lifetime of the controller to the lifetime of a
request object. Thus it can be used more freely in the application and
in test contexts.