Refactor RegisterHoursWorkedController
This commit changes the RegisterHoursWorkedController with regards to
how it receives the current request object. Before this commit the
request object was injected into the controller via the `__init__`
method. This limited the reusability of the class because it tied its
lifetime artificially to the lifetime of a request object. After this
change the controller receives the request object via a paramter of
the `create_use_case_request` method.