Refactor SelectProductiveConsumptionController
This commit changes how the SelectProductiveConsumptionController
receives request objects. Before this commit instances of the
controller were receiving the request objects via the `__init__`
method. This approach limited reusability of this class in production
and testing scenarios. After this commit the controller receives
request objects through a parameter of `process_input_data`.