Refactor CreateDraftFromPlanPresenter
This change refactors the CreateDraftFromPlanPresenter. Before this
change the presenter in question would receive the currently processed
request object via a paramter to its `__init__` function. After this
change the request object is passed as an argument to the
`render_response` method of the presenter. This refactoring allows for
greater code reuse as the lifetime of presenter objects now
independent of the lifetime of a request.