Refactor LoginMemberApiController
This commit changes how the LoginMemberApiController receives requests
objects to process. Before this change the request object that the
controller is supposed to process was passed to the controller via a
parameter of the controllers `__init__` method. Now the request object
is passed into the controller via a paramter of the `create_request`
method.