Replace Flask-Migrate with alembic
Before this commit our database migrations would depend on flask_migrate which depends on flask_sqlalchemy. We delete this dependency and are using plain alembic now.

We still have a slight dependency between alembic and Flask: In `arbeitszeit_flask.migrations.env.py` we make use of a Flask app object to retrieve the correct db uri. This is a dependency we can get rid of in the near future.

The github job has been adjusted to use alembic instead of Flask-Migrate.