New web form for draft creation and draft edits
We delete the WTForms implementation of the `CreateDraftForm` in `arbeitszeit_flask/forms.py` and use a form in `arbeitszeit_web/forms.py` instead, that is a simple python class. In the html code in `draft_form.html`, we now only define the `type` attribute of the input element (text, textarea or number). All other restrictions (min/max value, required, whole number, etc) are now implemented in the controller.

Apart from being less "magical" and better understandable, this new implementation allows better testing (without the need of testing the restrictions on flask level in integration tests) and more flexible error messages.

The downside is, that we have to implement the validation logic by ourself. This is done in the new `DraftFormValidator`.

Fixes #945