Filter SAWarning in pytest report
When running our tests via pytest, three tests in our testsuite show "sqlalchemy.exc.SAWarning"  in the test report. This is because we are trying to create a second email address with the same primary key. This commit filters the warnings.

The filtering happens on testcase level. The alternative, filtering via an entry in the `pyproject.toml` file was rejected as being considered too general.