Fix bug in auto migration
Before this commit, the AUTO_MIGRATE config option did not work as expected for the test or the production db, because our migration tool alembic would migrate always the the dev db. Now the correct db engine is passed to the migration script in `arbeitszeit_flask.migrations.env.py`, to make sure that the correct db is auto migrated.

Moreover, when auto migrating, alembic now "stamps" databases that were not initialized with alembic, to enable the migration on that db.