Do not stamp db to "head"
Do not stamp databases with alembic version of “head” when auto migrating. This leads to skipping of all missing migrations in the respective database, because Alembic sees the database as already up-to-date and does not upgrade.

Instead only create the alembic version table if it does not exist. This ensures that databases without alembic version table can get migrated as well.