Create TransferType enum
This class enumerates all possible transfer types. The names of the indiviual types are intended to be as short and comprehensible as possible. E.g., I chose the name `taxes` over more correct, but less  comprehensible alternatives.

The possible transfer types are deducted from the discussion in "RFC - Design of accounts and transactions" #1175

fixes #1198