Add tests for public sector fund
This commit adds tests for the public sector fund. They are testing the psf balance using abstract business logic (use cases and services) instead of transactions directly, which is more stable and reliable. It will show its merits when we model "wages" and "taxes" as transfers instead of transactions.

A new method of the `PlanGenerator` class has been created for these tests. The `create_plans_that_lead_to_fic_of` method takes the desired FIC and creates the necessary plans to achieve this.