Add tests for public sector fund
This commit adds tests for the public sector fund. They are testing the psf balance using abstract business logic (use cases and services) instead of transactions directly, which is more stable and reliable. This abstract testing approach will be necessary when we model wages and taxes as transfers instead of transactions.