Remove redundant plan.activation_date column
The activation_date column is redundant with plan_review.approval_date and is therefore not required. We always considered "activated" a synonym for "approved". This commit deletes the column from the database model "Plan". In the ApprovePlanUseCase, now only approval date is set, no setting of activation date is needed anymore. Function and variable names have been adapted.

fixes #1216