Add "type" field to transfer record
Transfers of labour time can only be of certain, limited types. It therefore makes sense to add an enum field "type" to the transfer record in order to distinguish between different kind of transfers without the need of elaborated database joins, etc. In the future, with arrival of undo-transfers or other new transfer types, it may be even impossible to distinguish transfers without this new field.

fixes #1219