Improve test isolation and performance of db and flask tests
This commit adds a new base test class, "DatabaseTestCase", which is intended to be used for database tests that do not need access to the flask app.

Furthermore the commit enhances the test isolation of integration tests by implementing a connection-level transaction in the new base test class, which gets rolled back in the `tearDown` method.

While all changes to the database in the tests are rolled back, the table creation, that happens in the create_app function, does not get rolled back. Keeping the tables between test runs allows for considerable better performance compared to before, where we dropped all tables after each test to achieve test isolation. On the authors machine, the `pytest` command now takes only 1:10 instead of 4:10 minutes.