Include Cooperation as potential AccountOwner
Added Cooperation to the AccountOwner type in arbeitszeit/records.py. Repository methods have been adjusted. Note that we do not expect cooperations to be receivers of transactions, even though account owners in general can be receivers. This is fine, becau
se in the current setting this cannot occur, and transactions will be history, soon.