Allow unlimited overdraw of member accounts
After this change, by setting the variable ALLOWED_OVERDRAW_MEMBER to `-1`, the allowed overdraw of member accounts can be configured to be unlimited.

This is useful for tests and theoretical scenarios, but also for very permissive productive scenarios.

Note that this decision practically prevents us to allow negative overdraw values other than `-1`, which the author of this commits finds acceptable, because such a feature would mean to prohibit the spending of existing labour time tokens.