Provide dev server with data generation CLI
With this commit we enhance the flask dev server by registering a data generation CLI on it. This CLI uses internally the data generators from the `tests` folder, which in turn use the domain logic (use cases), to create records like companies or plans in the development database. By using the domain logic we make sure to only create consistent data.

Run `flask generate --help` for more information.

The data generation CLI code is located entirely in the tests folder and thus cannot get accidentally rolled out and used in the production environment.