Simplify usage of Paginator
By passing the web request object to the Paginator, we were able to simplify the usage of this class.

This commit is based on the design decision that 1) presenters in arbeitszeit_web/www_presenters may expect a http web request protocol class (this is a slight deviation from the clean code architecture concept that assigns the presenter only the task to work with use case output) and 2) that the Paginator class expects a http web request.

Presenters and the paginator class can therefore be considered http-specfic. The partial effort to clean presenters from web requests is abandoned. The alternative would be to pass transport-specific information through the use case, which would be a larger deviation from clean architecture principles.