Stricter boundaries for list transfer use case
Instead of passing business objects (entities like Member or Company) from the ListTransfer use case to the presenter, pass specialized data transfer objects.

The downside is more parsing and mapping necessary, the upside is clearer boundaries between presenter and business logic.

Anonymization of workers' names, ids and account ids is now done in the use case instead of the presenter.