Fix behaviour of plan repository method implementations
The plan result method "that_are_in_same_cooperation_as(plan)" did mistakenly return the plan from the argument if that plan was not part of any cooperation.

The intended behaviour is decribed in a docstring of the method in `arbeitszeit/repositories.py` which states that the method should return None in that case.

The wrong behaviour had to be corrected both in the test and the flask implementation of the repository. Tests have been added.