Expand AccountDetailsService
With this commit, the AccountDetailsService returns information on the transfer counter party. If the requested account is the debit account, the counter party is the creditor, and vice versa.

With this change, the service is equipped to be used by all account overview interactors in the same way, which reduces code duplication and enhances maintainability.