Dev: Set default values in development settings
The logic in our "development_settings.py" had set some variables to None if they were not defined by the user. Consequently this "None" would override the values defined in "configuration_base.py".

This fixes a bug in our development server which would crash in certain views if account deviation (ACCEPTABLE_RELATIVE_ACCOUNT_DEVIATION) was not set by the developer.