Introduce ConfigValidator for flask config
Before this change we had no validation if all required options for the app (as specified in docs on "Hosting") are actually imported and successfully stored in flask's app config object.

This commit implements that validation, which happens on flask startup. The validation checks for existance and correct type. If validation fails, an exception is raised preventing the start of the app. This is preferable to have hard-to-detect unexpected app behaviour afterwards.

We expect the required options to be provided in all environments (test, dev, prod).

Config-related code has been moved into a new folder "arbeitszeit_flask/config"