Make colors customizable
After this commit, we can change the main colors of the flask app via CSS variables defined in a CSS file and in "arbeitszeit_flask/flask_colors.py". In the adapters layer nothing has changed, we can still access the main colors through the interface in arbeitszeit_web/colors.py.

The actual colors of the app have not changed.